requirements.txt
streamlit==1.44.1
firebase_admin==6.7.0
pillow==11.2.1
protobuf==6.30.2
pytz==2024.1
streamlit==1.44.1
google-generativeai==0.8.4
dotenv==0.9.9